replaced jest with mocha because mocha is much faster and jest is just bloated
