+++ package.json
@@ -14,11 +14,11 @@
     "cloudformation"
   ],
   "scripts": {
-    "test": "jest --runInBand --ci",
+    "test": "mocha test/*.test.js",
     "lint": "eslint dist/cdk-web.js",
     "build": "npm-run-all -s build:* lint",
     "clean": "git clean -xdf -e node_modules",
-    "coverage": "nyc report --reporter=html --compact",
+    "coverage": "CDK_WEB_DEBUG=1 npm test",
     "build:bundle": "webpack --config cdk-web.webpack.config.js",
     "build:verify": "tsc --noEmit index.d.ts",
     "build:compress": "google-closure-compiler --js=dist/cdk-web.js --js_output_file=dist/cdk-web.min.js --warning_level=QUIET",
@@ -67,22 +67,19 @@
     "glob": "^7.2.0",
     "google-closure-compiler": "^20220405.0.0",
     "ignore": "^5.2.0",
-    "jest": "^27.4.7",
-    "jest-puppeteer": "^6.1.0",
     "jsdoc-to-markdown": "^7.1.1",
     "jszip": "^3.9.1",
     "loader-utils": "^2.0.2",
     "lodash": "^4.17.21",
     "memfs": "^3.4.1",
+    "mocha": "^9.2.2",
     "npm-dts": "^1.3.11",
     "npm-run-all": "^4.1.5",
-    "nyc": "^15.1.0",
     "os-browserify": "^0.3.0",
     "path-browserify": "^1.0.1",
     "precinct": "^9.0.0",
     "process": "^0.11.10",
     "puppeteer": "^13.5.2",
-    "puppeteer-to-istanbul": "^1.4.0",
     "react": "~16",
     "react-bootstrap": "^2.1.2",
     "react-dom": "~16",
@@ -101,15 +98,10 @@
     "webpack-dev-server": "~3",
     "webpack-mild-compile": "^3.4.0"
   },
-  "jest": {
+  "mocha": {
     "bail": true,
-    "verbose": true,
-    "preset": "jest-puppeteer",
-    "testTimeout": 120000,
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "<rootDir>/docs"
-    ]
+    "timeout": 120000,
+    "file": "./test/_setup.js"
   },
   "browserslist": {
     "production": [
